New attribute `query_args` will pass a JSON string to `get_posts()`.
This makes it possible to use all the various arguments to `get_posts()`
from the shortcode in order to do things like filter out sepcific posts,
return the posts in a particular sort order, or anything else you can do
with the WP core `get_posts()` function.

Authors will need to write this value as a JSON-formatted string for it
to be correctly understood. For instance:

    [site-directory display="list" query_args='{"order": "ASC"}']

Note the single quoted value to the `query_args` attribute.

git-svn-id: https://plugins.svn.wordpress.org/multisite-directory/trunk@1393790 b8457f37-d9ea-0310-8a92-e5e31aec5664