Fixes #13: Remove sites from the directory when the site is deleted.
This fixes the shortcode tests so that they are no longer dependent on
when in the test sequence they were run. Previously to this, the
shortcode's unit tests had to be run first or they would fail, because
the tests were written to expect specific blog ID values. These are
created by MySQL's `auto_increment` property and, as a result, would
fail when a site was created or deleted before the shortcode tests ran.

git-svn-id: https://plugins.svn.wordpress.org/multisite-directory/trunk@1395031 b8457f37-d9ea-0310-8a92-e5e31aec5664