Use static functions so these hooks can be removed from other code.
Using static functions allows other code to access and modify these
hooks. For example, a different plugin or a theme's `functions.php` file
can now write code such as

    remove_action('subsite_category_edit_form_fields', array('Multisite_Directory_Taxonomy', 'edit_form_fields));

to remove the map from the subsite category's editing interface. Prior
to this change, these were never removable because the callable was
bound to `$this`, which couldn't be referenced from outside the class.

See also
https://wordpress.org/support/topic/hiding-the-geo-parts?replies=2#post-8338275

git-svn-id: https://plugins.svn.wordpress.org/multisite-directory/trunk@1406109 b8457f37-d9ea-0310-8a92-e5e31aec5664